% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{addData}
\alias{addData}
\title{Add data to an initialized CytoFlag object}
\usage{
addData(CF, input, slot, n = 1000)
}
\arguments{
\item{CF}{CytoFlag object}

\item{input}{List of FCS file paths}

\item{slot}{Slot in the CytoFlag object to store data}

\item{n}{Total number of cells to read per file (default = 1000)}
}
\value{
CytoFlag object
}
\description{
Add data to an initialized CytoFlag object
}
