% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{addReferencelabels}
\alias{addReferencelabels}
\title{Add additional annotations for reference data}
\usage{
addReferencelabels(CF, labels)
}
\arguments{
\item{CF}{CytoFlag object}

\item{labels}{Vector of numeric or string labels}
}
\value{
CytoFlag object
}
\description{
Add additional annotations for reference data
}
\seealso{
\code{\link{addTestlabels}}
}
