% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{addTestdata}
\alias{addTestdata}
\title{Add reference data to initialized CytoFlag object}
\usage{
addTestdata(CF, input, read = FALSE, reload = FALSE, aggSize = 10000)
}
\arguments{
\item{CF}{CytoFlag object}

\item{input}{List of FCS file paths}

\item{read}{Whether to read FCS files already (default = FALSE)}

\item{reload}{Whether to read ALL fcs files, including previously loaded (default = FALSE)}

\item{aggSize}{How many cells to use in total for aggregate sample (default = 10000)}
}
\value{
CytoFlag object
}
\description{
Add reference data to initialized CytoFlag object
}
