% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{addTestlabels}
\alias{addTestlabels}
\title{Add additional annotations for test data}
\usage{
addTestlabels(CF, labels)
}
\arguments{
\item{CF}{CytoFlag object}

\item{labels}{Vector of numeric or string labels}
}
\value{
CytoFlag object
}
\description{
Add additional annotations for test data
}
\seealso{
\code{\link{addReferencelabels}}
}
