% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{CytoFlag}
\alias{CytoFlag}
\title{Run CytoFlag.}
\usage{
CytoFlag(
  input,
  channels,
  compensate = TRUE,
  transform = TRUE,
  transformFunction = flowCore::logicleTransform(),
  aggSize = 10000,
  featMethod = "summary",
  PCA = TRUE,
  flagMethod = "forest"
)
}
\arguments{
\item{input}{An array of paths to files or directories}

\item{channels}{An array of channels to use for anomaly detection}

\item{compensate}{logical, does the data need to be compensated}

\item{transform}{logical, does the data need to be transformed with the
transformation given in \code{transformFunction}}

\item{transformFunction}{Defaults to logicleTransform()}

\item{aggSize}{integer, size of aggregated flowframe for model training}

\item{featMethod}{character, feature generation method to use}

\item{PCA}{logical, reduce features to first 2 principal components}

\item{flagMethod}{character, algorithm to use for anomaly detection}

\item{transformList}{transformList to apply on the samples}
}
\value{
Generated features and outlier label for \code{input}
}
\description{
Function to run CytoFlag workflow. Will automatically remove margin events,
compensate and transform FCS files.
}
\examples{
CytoFlag(list.files(FCS_dir), featMethod = "summary", flagMethod = "forest")
CytoFlag(list(list.files(reference_dir), list.files(test_dir),
         featMethod = "summary", flagMethod = "forest")

}
