% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFeatures.R
\name{generateFeatures}
\alias{generateFeatures}
\title{Generate features and attach to CytoFlag object}
\usage{
generateFeatures(
  CF,
  channels,
  featMethod = "summary",
  n = 1000,
  aggSize = 10000,
  aggSlot = "auto",
  cores = "auto",
  recalculate = FALSE,
  nRecursions = 4
)
}
\arguments{
\item{CF}{CytoFlag object}

\item{channels}{Channels to calculate features for}

\item{featMethod}{Feature generation method to use}

\item{n}{How many cells to use for feature generation}

\item{aggSize}{How many cells to use in total for aggregate sample (default = 10000)}

\item{aggSlot}{Whether to use reference or test data as aggregate (default = "auto")}

\item{cores}{How many cores to use for parallelization (default = 50%)}

\item{recalculate}{Whether to recalculate features for existing data}

\item{nRecursions}{Number of recursions to use for fingerprinting (default = 4)}
}
\value{
Dataframe with features (columns) for samples (rows)
}
\description{
Generate features and attach to CytoFlag object
}
