% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(
  CF,
  featMethod,
  plotRef = FALSE,
  flagSlot = NULL,
  plotLabels = FALSE
)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Which generated features to plot}

\item{plotRef}{Whether to visualize reference, if available (default = TRUE)}

\item{flagSlot}{Flags to plot (NULL, "outliers", "novelties") (default = NULL)}

\item{plotLabels}{Whether to plot supplied labels}
}
\value{
Heatmap plot
}
\description{
Plot Heatmap
}
