% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Plot PCA}
\usage{
PlotPCA(
  CF,
  featMethod,
  plotRef = FALSE,
  flagSlot = NULL,
  ids = FALSE,
  plotLabels = FALSE
)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Which generated features to plot}

\item{plotRef}{Whether to visualize reference, if available (default = TRUE)}

\item{flagSlot}{Flags to plot (NULL, "outliers", "novelties") (default = NULL)}

\item{ids}{Whether to plot indices of anomalies (default = FALSE)}

\item{plotLabels}{Whether to plot supplied labels, overrides flagSlot}
}
\value{
PCA plot
}
\description{
Plot PCA
}
