% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{readInput}
\alias{readInput}
\title{Read, preprocess and downsample FCS files}
\usage{
readInput(CF, path, n = NULL)
}
\arguments{
\item{CF}{CytoFlag object}

\item{path}{Location of FCS file}

\item{n}{Total number of cells to read per file}
}
\value{
flowFrame
}
\description{
Read, preprocess and downsample FCS files
}
