% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDim.R
\name{reduceDim}
\alias{reduceDim}
\title{Reduce features to first two principal components}
\usage{
reduceDim(testFeatures, flagStrat, refFeatures = NULL)
}
\arguments{
\item{testFeatures}{Dataframe of features from test samples}

\item{flagStrat}{Which anomaly detection strategy to use ("outlier" or "novelty").}

\item{refFeatures}{Dataframe of features from reference samples}
}
\value{
List containing PC loadings, variance and reduced features
}
\description{
Reduce features to first two principal components
}
