% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{CytoFlag}
\alias{CytoFlag}
\title{Initialize CytoFlag object}
\usage{
CytoFlag()
}
\value{
Default initialized CytoFlag object
}
\description{
Function to initialize CytoFlag workflow. Returns CytoFlab object
}
