% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{Flag}
\alias{Flag}
\title{Flag anomalies in CytoFlag object}
\usage{
Flag(CF, featMethod = NULL, flagStrat = "outlier", PCA = FALSE)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Feature generation method to use.
Either one of: "summary", "EMD", "binning" or "fingerprint".}

\item{flagStrat}{Which anomaly detection strategy to use ("outlier" or "novelty").
By default, selects intended use based on availability of reference material.}

\item{PCA}{Whether to reduce dimensionality to first two PCs. (default = FALSE)}
}
\value{
CytoFlag object
}
\description{
Flag anomalies in CytoFlag object
}
