% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{Flag}
\alias{Flag}
\title{Flag anomalies based on generated features}
\usage{
Flag(CF, featMethod, flagStrat, PCA = FALSE)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Which features to use for anomaly detection}

\item{flagStrat}{Which flaggins strategy to use ("outlier" or "novelty")}

\item{PCA}{Reduce features to 2D PCA before flagging (default = FALSE)}
}
\value{
CytoFlag object
}
\description{
Flag anomalies based on generated features
}
\seealso{
\code{\link{generateFeatures}}
}
