% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoFlag.R
\name{Flag}
\alias{Flag}
\title{Flag anomalies based on generated features}
\usage{
Flag(CF, featMethod, flagMethod)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Which features to use for anomaly detection}

\item{flagMethod}{Which type of anomaly detection to use ("outlier" or "novelty")}
}
\value{
CytoFlag object
}
\description{
Flag anomalies based on generated features
}
\seealso{
\code{\link{generateFeatures}}
}
