% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(
  CF,
  featMethod,
  fitData,
  plotData,
  PCx = 1,
  PCy = 2,
  color = NULL,
  shape = NULL,
  plotLoadings = TRUE
)
}
\arguments{
\item{CF}{CytoFlag object}

\item{featMethod}{Which features to use}

\item{fitData}{Which data to fit PCA on ("test", "reference" or "all")}

\item{plotData}{Which data to plot based on PCA fit ("test", "reference" or "all")}

\item{PCx}{Which PC to plot along x-axis}

\item{PCy}{Which PC to plot along y-axis}

\item{color}{Category used for point color ("outlier", "novelty", "labels")}

\item{shape}{Category used for point shape ("outlier", "novelty", "labels")}

\item{plotLoadings}{Whether to plot loadings along x and y-axis}
}
\value{
PCA plot
}
\description{
Plot PCA
}
